Move CustomOAuth2UserService.java into service folder
